+++ src/main/kotlin/com/github/siropkin/kursor/KursorSettings.kt
@@ -4,12 +4,10 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
-import com.intellij.openapi.editor.colors.EditorColorsManager
 import com.intellij.util.xmlb.Converter
 import com.intellij.util.xmlb.XmlSerializerUtil
 import com.intellij.util.xmlb.annotations.OptionTag
 import java.awt.Color
-import java.awt.Font
 
 
 @State(
@@ -19,22 +17,18 @@ import java.awt.Font
 class KursorSettings : PersistentStateComponent<KursorSettings> {
     var defaultLanguage: String = "us"
 
-    var changeColorOnNonDefaultLanguage: Boolean = true
-    @OptionTag("colorOnNonDefaultLanguage_", converter = ColorConverter::class)
-    var colorOnNonDefaultLanguage: Color = Color(255, 140, 0)
+    @OptionTag("cursorColor_", converter = NullableColorConverter::class)
+    var cursorColor: Color? = Color(255, 140, 0)
 
     var showTextIndicator: Boolean = true
-
-    var textIndicatorFontName: String = EditorColorsManager.getInstance().globalScheme.editorFontName
-    var textIndicatorFontStyle: Int = Font.PLAIN
-    var textIndicatorFontSize: Int = 11
-    var textIndicatorFontAlpha: Int = 180
-
-    var textIndicatorVerticalPosition: String = TextIndicatorVerticalPositions.TOP
-    var textIndicatorHorizontalOffset: Int = 4
-
     var indicateCapsLock: Boolean = true
     var indicateDefaultLanguage: Boolean = false
+
+    @OptionTag("textIndicatorColor_", converter = ColorConverter::class)
+    var textIndicatorColor: Color = Color(255, 140, 0)
+
+    @OptionTag("textIndicatorBackgroundColor_", converter = NullableColorConverter::class)
+    var textIndicatorBackgroundColor: Color? = null
 
     override fun getState(): KursorSettings {
         return this
@@ -58,10 +52,23 @@ class KursorSettings : PersistentStateComponent<KursorSettings> {
 class ColorConverter : Converter<Color>() {
     override fun fromString(value: String): Color {
         val parts = value.split(",")
-        return Color(parts[0].toInt(), parts[1].toInt(), parts[2].toInt(), parts[3].toInt())
+        return Color(parts[0].toInt(), parts[1].toInt(), parts[2].toInt())
     }
 
     override fun toString(value: Color): String {
-        return "${value.red},${value.green},${value.blue},${value.alpha}"
+        return "${value.red},${value.green},${value.blue}"
+    }
+}
+
+class NullableColorConverter : Converter<Color?>() {
+    override fun fromString(value: String): Color? {
+        if (value.isBlank()) return null
+        val parts = value.split(",")
+        return Color(parts[0].toInt(), parts[1].toInt(), parts[2].toInt())
+    }
+
+    override fun toString(value: Color?): String {
+        if (value == null) return ""
+        return "${value.red},${value.green},${value.blue}"
     }
 }